feat: group components by screen
