+++ src/screens/Partnership/MeetingsList/index.tsx
@@ -1,0 +1,5 @@
+import { Text } from "@components";
+
+export function MeetingsList() {
+  return <Text>Em breve...</Text>;
+}
