+++ src/screens/Partnership/index.tsx
@@ -1,32 +1,19 @@
-import {
-  Button,
-  Card,
-  Header,
-  Input,
-  Loading,
-  Modal,
-  Tabs,
-  Text,
-} from "@components";
+import { Button, Header, Tabs, Text } from "@components";
 import { RootStackParamList } from "@custom-types/rootStackParamList";
-import { IComment } from "@interfaces/annotation.interface";
 import { IPartnership } from "@interfaces/partner.interface";
 import { RouteProp, useRoute } from "@react-navigation/native";
-import annotationRequests from "@requests/annotation.requests";
 import partnershipRequests from "@requests/partnership.requests";
-import { formatDate } from "@utils/formatDate";
-import { formatTime } from "@utils/formatTime";
 import { useEffect, useState } from "react";
 import { ScrollView, View } from "react-native";
+import { AnnotationsList } from "./AnnotationsList";
 import { EditPartnershipModal } from "./EditPartnershipModal";
+import { MeetingsList } from "./MeetingsList";
 import {
   ButtonsContainer,
   ContactView,
   Container,
   HistoryContainer,
   InformationView,
-  ListContainer,
-  LoadingContainer,
   PartnerInfoView,
 } from "./styles";
 
@@ -60,10 +47,6 @@ export function Partnership() {
     getPartnerships();
   }
 
-  async function handleUpdatePartnership() {
-    setVisibleEditModal(true);
-  }
-
   return (
     <Container>
       <Header />
@@ -88,7 +71,7 @@ export function Partnership() {
               Deletar parceria
             </Button>
             <Button
-              onPress={handleUpdatePartnership}
+              onPress={() => setVisibleEditModal(true)}
               style={{ marginVertical: 8 }}
             >
               Editar informações
@@ -156,149 +139,15 @@ export function Partnership() {
           <HistoryContainer>
             <Tabs onChangeTab={tab => setTab(tab)} />
             {tab === 0 ? (
-              <History isDisabled={isLoading || (data?.disabled ?? false)} />
+              <AnnotationsList
+                isPartnershipDisabled={isLoading || (data?.disabled ?? false)}
+              />
             ) : (
-              <MeetingList />
+              <MeetingsList />
             )}
           </HistoryContainer>
         </ScrollView>
       )}
     </Container>
   );
-}
-
-interface HistoryProps {
-  isDisabled: boolean;
-}
-
-function History({ isDisabled }: HistoryProps) {
-  const [newComment, setNewComment] = useState("");
-  const [isLoading, setIsLoading] = useState(true);
-  const [annotations, setAnnotations] = useState<IComment[]>();
-  const [modalComment, setModalComment] = useState<IComment>();
-  const [isCommentModalOpen, setIsCommentModalOpen] = useState(false);
-  const [isEditCommentModalOpen, setIsEditCommentModalOpen] = useState(false);
-  const [editedComment, setEditedComment] = useState("");
-  const route = useRoute<RouteProp<RootStackParamList, "Partnership">>();
-
-  const { id } = route.params;
-
-  async function getData() {
-    setIsLoading(true);
-    const comments = await annotationRequests.getAnnotations(id);
-    setAnnotations(comments);
-    setIsLoading(false);
-  }
-
-  useEffect(() => {
-    getData();
-  }, [id]);
-
-  async function handleAddComment() {
-    setIsLoading(true);
-    await annotationRequests.createAnnotation(id, newComment);
-    const updatedComments = await annotationRequests.getAnnotations(id);
-    updatedComments && setAnnotations(updatedComments);
-    setNewComment("");
-    setIsLoading(false);
-  }
-
-  async function handleEditComment() {
-    setIsLoading(true);
-    await annotationRequests.updateAnnotation(
-      modalComment?.id ?? "",
-      id,
-      editedComment,
-    );
-    const updatedComments = await annotationRequests.getAnnotations(id);
-    updatedComments && setAnnotations(updatedComments);
-    setEditedComment("");
-    setIsLoading(false);
-    setIsEditCommentModalOpen(false);
-  }
-
-  return (
-    <ListContainer scrollEnabled>
-      {!isDisabled && (
-        <Input
-          label={"Inserir atualização"}
-          placeholder={"Nova atualização sobre a parceria..."}
-          value={newComment}
-          onChangeText={text => setNewComment(text)}
-          hasOutIcon
-          onPressIcon={handleAddComment}
-        />
-      )}
-
-      <Text size={14} color={"#666666"} style={{ marginVertical: 16 }}>
-        Todas as atualizações e anotações
-      </Text>
-
-      {isLoading ? (
-        <LoadingContainer>
-          <Loading />
-        </LoadingContainer>
-      ) : annotations?.length === 0 ? (
-        <Text
-          size={14}
-          color={"#999999"}
-          style={{ textAlign: "center", marginVertical: 24 }}
-        >
-          Sem atualizações ou anotações
-        </Text>
-      ) : (
-        annotations?.map(card => {
-          const isEdited = card.createdAt != card.updatedAt;
-
-          return (
-            <Card
-              key={card.id}
-              id={card.id}
-              type={card.title ? "annotation" : "update"}
-              date={formatDate(isEdited ? card.updatedAt : card.createdAt)}
-              time={formatTime(isEdited ? card.updatedAt : card.createdAt)}
-              description={card.comment}
-              author={`${card.User.name} ${card.User.lastName}`}
-              title={card.title}
-              onPress={() => {
-                setModalComment(card);
-                setIsCommentModalOpen(true);
-              }}
-              onEdit={() => {
-                setModalComment(card);
-                setEditedComment(card.comment);
-                setIsEditCommentModalOpen(true);
-              }}
-            />
-          );
-        })
-      )}
-
-      <Modal
-        visible={isCommentModalOpen}
-        onClose={() => setIsCommentModalOpen(false)}
-        title={`Autor(a): ${modalComment?.User.name}`}
-        content={<Text>{modalComment?.comment}</Text>}
-      />
-
-      <Modal
-        visible={isEditCommentModalOpen}
-        onClose={() => setIsEditCommentModalOpen(false)}
-        title={"Editar comentário"}
-        buttonTitle="Editar comentário"
-        onPressButton={handleEditComment}
-        isLoading={isLoading}
-        content={
-          <Input
-            value={editedComment}
-            onChangeText={text => setEditedComment(text)}
-          />
-        }
-      />
-    </ListContainer>
-  );
-}
-
-function MeetingList() {
-  return <Text>Em breve...</Text>;
 }
