+++ src/screens/Partnership/styles.ts
@@ -19,12 +19,6 @@ export const HistoryContainer = styled.View`
   flex: 1;
 `;
 
-export const ListContainer = styled.ScrollView``;
-
-export const LoadingContainer = styled.View`
-  margin-top: 24px;
-`;
-
 export const TextInput = styled.TextInput`
   background: #f4f5f7;
   padding: 0px 16px;
